feat - Moved divs inside 'AboutDiv' and switching them instead
Does not hide the div in the content div, instead is now hiding the third column
inside that div and the changing content is in there. The picture and the brief
about is consistent.

Some aspects do not take a lot of space, hence it is better to just keep the
image and the about text and just swap the skills section with the contact and
the doc section.